Add non canonical mode during user input
In order to process command keys such as direction keys, one need to
have read return as soon as input is available, without waiting for a
complete line. Which is the purpose of non canonical mode.
Additionally, the ECHOCTL flag is diasable in order not to echo the
escape sequences corresponding to those commands when the get them.