First attempt for cmd keys recognition
Try to identify sequences matching the various command keys sequences
that may be sent by the terminal (such as the arrows keys).
Reading is done one character at a time. Doing otherwise would cause
some trouble, like waiting for the read to complete when there is
actually nothing more to read.
(There might be workaround for that, but this is not the path choosed
for now).
The assumptions which are made about those sequences are the folowing :
- no sequence is a subset or another one
Follow-up : two or more sequences can always be discrimined by
requesting another character.