Add a new object : term_device
The object line_editor has two essentials part : a buffer to hold user
entry, and a term_device which has to stay updated with the buffer. So
it makes sense to divide it into these two objects.

Separation is not yet fully done : the functions to search the key
sequences use both headers. To be done.