Correct destroy for term_device
Assumptions were made that tgetstr was using malloc and never freeing.
as documented in the termcap documentation.
However, it appears that this is not accurate (trying to free these
adresse raises a free-on-non-malloced-memory error). This is probably
due to the fact that those the termcap functions are in fact implemented
by the ncurses library.