workaround: Delete at end of line
Whenever I try to delete a character at end of line (being on the line
below after a wrap, it does not work. Reducing the number of column the
shell see fix the problem, for now. But this it not really a clean way
to go, and there will be work to do about this later.